# docs/swagger.yaml
definitions:
  auth.LoginRequest:
    properties:
      email:
        example: john.doe@example.com
        type: string
      password:
        example: SecurePass123
        type: string
    required:
    - email
    - password
    type: object
  auth.LoginResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      expires_in:
        example: 3600
        type: integer
      token_type:
        example: Bearer
        type: string
      user:
        $ref: '#/definitions/auth.UserResponse'
    type: object
  auth.RegisterRequest:
    properties:
      email:
        example: john.doe@example.com
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: SecurePass123
        type: string
      role:
        description: Optional, only allowed for admin users
        example: user
        type: string
    required:
    - email
    - name
    - password
    type: object
  auth.RegisterResponse:
    properties:
      message:
        example: Registration successful. Welcome email has been sent.
        type: string
      user:
        $ref: '#/definitions/auth.UserResponse'
    type: object
  auth.UserResponse:
    properties:
      created_at:
        example: "2025-07-19T10:30:00Z"
        type: string
      email:
        example: john.doe@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: John Doe
        type: string
      role:
        example: user
        type: string
      updated_at:
        example: "2025-07-19T10:30:00Z"
        type: string
    type: object
  auth.ValidationError:
    properties:
      field:
        example: email
        type: string
      message:
        example: email format is invalid
        type: string
    type: object
  auth.ValidationErrors:
    properties:
      errors:
        items:
          $ref: '#/definitions/auth.ValidationError'
        type: array
    type: object
  common.BaseResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: User login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            allOf:
            - $ref: '#/definitions/common.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.LoginResponse'
              type: object
        "400":
          description: Bad request - validation errors
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "401":
          description: Unauthorized - invalid credentials
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "422":
          description: Validation failed
          schema:
            allOf:
            - $ref: '#/definitions/common.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.ValidationErrors'
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      summary: User login
      tags:
      - Authentication
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account with email and password. Admins can
        specify role.
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            allOf:
            - $ref: '#/definitions/common.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.RegisterResponse'
              type: object
        "400":
          description: Bad request - validation errors
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "401":
          description: Unauthorized - invalid token
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "409":
          description: Conflict - email already exists
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "422":
          description: Validation failed
          schema:
            allOf:
            - $ref: '#/definitions/common.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.ValidationErrors'
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Register a new user
      tags:
      - Authentication
...
swagger: "2.0"